Figured out issue...
The issue was that .from_csv works with repositories because
repos take a table of data and pass it to populate repo, which
creates new children from a row of data. Currently from_csv only
accepts a path and returns a table. Could maybe work around...

It's either live with it or go back at this point.

Don't use .from_csv when invoking a single instance of a child.
Or, create a row object to pass in?
Or, write a separate child class methods module. The whole point of